move linode specific types to linode features
