+++ packages/manager/src/features/Images/ImagesCreate/CreateImageTab.tsx
@@ -33,7 +33,7 @@ import { useRegionsQuery } from 'src/queries/regions/regions';
 import { getQueryParamsFromQueryString } from 'src/utilities/queryParams';
 
 import type { CreateImagePayload } from '@linode/api-v4';
-import type { LinodeConfigAndDiskQueryParams } from 'src/utilities/queryParams';
+import type { LinodeConfigAndDiskQueryParams } from 'src/features/Linodes/types';
 
 export const CreateImageTab = () => {
   const location = useLocation();
