+++ packages/manager/src/features/Linodes/LinodesCreate/LinodeCreate.tsx
@@ -106,9 +106,9 @@ import type { RegionsProps } from 'src/containers/regions.container';
 import type { WithTypesProps } from 'src/containers/types.container';
 import type { WithLinodesProps } from 'src/containers/withLinodes.container';
 import type { LinodeCreateType } from 'src/features/Linodes/LinodesCreate/types';
+import type { LinodeCreateQueryParams } from 'src/features/Linodes/types';
 import type { CreateTypes } from 'src/store/linodeCreate/linodeCreate.actions';
 import type { ExtendedIP } from 'src/utilities/ipUtils';
-import type { LinodeCreateQueryParams } from 'src/utilities/queryParams';
 
 export interface LinodeCreateProps {
   additionalIPv4RangesForVPC: ExtendedIP[];
