+++ packages/manager/src/features/Linodes/LinodesDetail/LinodesDetail.tsx
@@ -14,7 +14,7 @@ import { SuspenseLoader } from 'src/components/SuspenseLoader';
 import { useLinodeQuery } from 'src/queries/linodes/linodes';
 import { getQueryParamsFromQueryString } from 'src/utilities/queryParams';
 
-import type { LinodeConfigAndDiskQueryParams } from 'src/utilities/queryParams';
+import type { LinodeConfigAndDiskQueryParams } from 'src/features/Linodes/types';
 
 const LinodesDetailHeader = React.lazy(
   () => import('./LinodesDetailHeader/LinodeDetailHeader')
