+++ packages/manager/src/features/Linodes/types.ts
@@ -1,3 +1,6 @@
+import type { LinodeCreateType } from 'src/features/Linodes/LinodesCreate/types';
+import type { BaseQueryParams } from 'src/utilities/queryParams';
+
 export type DialogType =
   | 'delete'
   | 'detach_vlan'
@@ -7,8 +10,20 @@ export type DialogType =
   | 'rescue'
   | 'resize'
   | 'upgrade_volumes';
+
 export type OpenDialog = (
   type: DialogType,
   linodeID: number,
   linodeLabel?: string
 ) => void;
+
+export interface LinodeCreateQueryParams extends BaseQueryParams {
+  type: LinodeCreateType;
+}
+
+export interface LinodeConfigAndDiskQueryParams extends BaseQueryParams {
+  selectedDisk: string;
+  selectedLinode: string;
+}
+
+export type BooleanString = 'false' | 'true';
